Fixed some Defects from Octane: 1. D18052 - After adding model global parameter getting unexpected error under BF tab
2. D17052 - External integrations changes isn't showing into modified solution item.

3. D16001 - Getting Star sign (Astrik mark) for Business Flows when we are coming back from Automate tab without making any changes

also fixed an issue on UCRadioButtons, the radio button is checked property was binded to an enum value, and its update trigger was OnPropertyChanged which I have added in previous commits. removed the redundant binding between the Radio button is checked property to the enum value.

also in order to fix D16001, turned dirty tracking off while we are loading the action edit page which would've mark the Action->Activtiy->BF->Runset as dirty.

fixed an issue where deleting all Model Global Parameters/Remote Services Grid would not set the CurrentSelectedItem to null and kept reference to the removed object.

with that said, it is requiring further testing on what happens when we remove all BF/Agents/any tree items